requirements.txt
kaggle
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn
Pillow
nltk
tensorflow-text
ipython
streamlit
Pillow